backend/requirements.txt
fastapi
uvicorn
pydantic
sqlite3
python-multipart
pandas
matplotlib
tabula-py
groq
